Fixed: sprites were completely screwed when partially behind left border of the screen

+++ src/data/Sprite16A.cpp
@@ -75,7 +75,7 @@ void Sprite16A::Draw(DrawingContext& ctx, int32_t x, int32_t y, uint32_t index, 
 			{
 				buffer += rleType & 0xFF;
 				inX += rleType & 0xFF;
-				if (inX >= frame.mWidth+x)
+				if (inX >= static_cast<int32_t>(frame.mWidth) + x)
 				{
 					inX = x + (inX - (frame.mWidth+x));
 					inY++;
